refactor(mail): 모듈명 수정 및 구조 일부 수정
