+++ hongit-api/src/main/kotlin/com/devfloor/hongit/api/auth/presentation/AuthController.kt
@@ -1,0 +1,3 @@
+package com.devfloor.hongit.api.auth.presentation
+
+class AuthController
