+++ hongit-core/src/main/kotlin/com/devfloor/hongit/core/authtoken/AuthToken.kt
@@ -14,15 +14,13 @@ import javax.persistence.Table
  * 이메일 인증 시 사용되는 토큰을 관리하는 entity
  *
  * @property id 아이디
- * @property userId 토큰 대상 유저 ID
  * @property expirationTime 토큰 만료 시간
  * @property expired 토큰 만료 상태
  */
 @Entity
 @Table(name = "auth_tokens")
 class AuthToken(
-    userId: Long,
-    expirationTime: LocalDateTime,
+    expirationTime: LocalDateTime = LocalDateTime.now().plusDays(TOKEN_EXPIRATION_DATE),
     expired: Boolean = false,
     id: UUID = UUID.randomUUID(),
 ) : BaseEntity() {
@@ -32,19 +30,11 @@ class AuthToken(
     @Column(name = "id", columnDefinition = "BINARY(16)")
     val id: UUID = id
 
-    @Column(name = "user_id")
-    val userId: Long = userId
-
     @Column(name = "expiration_time")
     val expirationTime: LocalDateTime = expirationTime
 
     @Column(name = "expired")
     var expired: Boolean = expired
-
-    constructor(userId: Long) : this(
-        userId = userId,
-        expirationTime = LocalDateTime.now().plusDays(TOKEN_EXPIRATION_DATE),
-    )
 
     fun useToken() {
         this.expired = true
