--- mail-auth-client/src/main/kotlin/com/devfloor/hongit/client/mail/common/config/MailAuthProperties.kt
@@ -1,10 +1,0 @@
-package com.devfloor.hongit.client.mail.common.config
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.context.properties.ConstructorBinding
-
-@ConstructorBinding
-@ConfigurationProperties(prefix = "hongit")
-class MailAuthProperties(
-    val authUrl: String,
-)
