--- mail-auth-client/src/main/kotlin/com/devfloor/hongit/client/mail/common/utils/Constants.kt
@@ -1,5 +1,0 @@
-package com.devfloor.hongit.client.mail.common.utils
-
-const val MAIL_TEMPLATE_FILE = "mail-template"
-
-const val MAIL_SUBJECT = "Hongit 회원가입 이메일 인증"
