--- mail-auth-client/src/main/resources/application.yml
@@ -1,35 +1,0 @@
-spring:
-  thymeleaf:
-    prefix: classpath:templates/
-    suffix: .html
-    cache: false
-
-mail:
-  sender:
-    host: smtp.gmail.com
-    port: 587
-    username: useremail
-    password: userpassword
-    protocol: smtp
-    auth: true
-    starttls-enable: true
-    starttls-required: true
-    debug: true
-    connection-timeout: 5000
-    timeout: 5000
-    write-timeout: 5000
-
-# 임시로 주소 설정
-hongit:
-  auth-url: http://localhost:8080/auth
-
----
-
-spring:
-  config:
-    activate:
-      on-profile: prod
-
-# 임시로 주소 설정
-hongit:
-  auth-url: https://www.hongit.co.kr/auth
