+++ mail-client/src/main/kotlin/com/devfloor/hongit/client/mail/MailConfig.kt
