+++ mail-client/src/main/kotlin/com/devfloor/hongit/client/mail/common/config/MailSenderProperties.kt
