+++ mail-client/src/main/kotlin/com/devfloor/hongit/client/mail/common/utils/Constants.kt
@@ -1,0 +1,7 @@
+package com.devfloor.hongit.client.mail.common.utils
+
+const val MAIL_TEMPLATE_FILE = "mail-template"
+
+const val SIGN_UP_URL = "https://hongit.co.kr/signup"
+
+const val MAIL_SUBJECT = "Hongit 회원가입 이메일 인증"
