+++ mail-client/src/main/kotlin/com/devfloor/hongit/client/mail/domain/spec/MailSender.kt
@@ -1,0 +1,12 @@
+package com.devfloor.hongit.client.mail.domain.spec
+
+import com.devfloor.hongit.client.mail.application.request.MailRequest
+import org.springframework.scheduling.annotation.Async
+import javax.mail.internet.MimeMessage
+
+interface MailSender {
+    @Async
+    fun send(message: MimeMessage)
+
+    fun createMimeMessage(request: MailRequest, tokenId: String): MimeMessage
+}
