+++ mail-client/src/main/resources/application.yml
@@ -1,0 +1,20 @@
+spring:
+  thymeleaf:
+    prefix: classpath:templates/
+    suffix: .html
+    cache: false
+
+mail:
+  sender:
+    host: smtp.gmail.com
+    port: 587
+    username: useremail
+    password: userpassword
+    protocol: smtp
+    auth: true
+    starttls-enable: true
+    starttls-required: true
+    debug: true
+    connection-timeout: 5000
+    timeout: 5000
+    write-timeout: 5000
