+++ mail-client/src/main/resources/templates/mail-template.html
